Restructure assistant module naming and imports
