+++ AGENTS.md
@@ -20,3 +20,9 @@ Local Testing
 - 첫 실행 시 테스트 컨테이너 이미지(MongoDB, Qdrant, Llama.cpp)가 자동으로 풀(Pull)됩니다.
 - 별도의 API 키 설정 없이 통합 테스트가 동작하도록 구성되어 있습니다.
 - 특정 테스트만 실행하려면 파일 경로 또는 패턴을 지정하세요. 예) `flox activate -- uv run pytest naraninyeo/tests/integration/test_new_message_handler.py -q`
+
+Code Style
+- `naraninyeo/assistant` 패키지의 모듈 간 import 시에는 반드시 절대 경로(`from naraninyeo.assistant...`)를 사용하세요.
+- `naraninyeo/assistant/__init__.py`는 얇은 상태를 유지하고, 필요한 심볼은 각 모듈에서 직접 import 해서 사용하세요.
+- `naraninyeo/assistant` 모듈들에 `__all__` 목록을 새로 추가하지 말고 필요한 이름을 직접 import 해 사용하세요.
+- `core`/`services` 같은 모호한 패키지 이름을 다시 추가하지 말고, 도메인을 설명하는 명확한 모듈명을 사용하세요.
